+++ rpctypes.h
@@ -48,11 +48,11 @@ struct rpcbuff* rpcbuff_create(uint64_t* dimsizes,uint64_t dimsizes_len);  /*las
                                                                                                 this  endpoint wasnt writen by rpcbuff_pushto), but rpcbuff_pushto will fit bigger data*/
 void rpcbuff_free_internals(struct rpcbuff* rpcbuff); //same as _rpcbuff_free but dont free struct rpcbuff* (only internals of it)
 void rpcbuff_free(struct rpcbuff* rpcbuff);
-int rpcbuff_getlast_from(struct  rpcbuff* rpcbuff, uint64_t* index, uint64_t index_len,void* otype,uint64_t* otype_len,enum rpctypes type);
-int rpcbuff_pushto(struct rpcbuff* rpcbuff, uint64_t* index, uint64_t index_len, void* ptype,uint64_t type_len,enum rpctypes type);
+int rpcbuff_getlast_from(struct  rpcbuff* rpcbuff, uint64_t* index, size_t index_len,void* otype,uint64_t* otype_len,enum rpctypes type);
+int rpcbuff_pushto(struct rpcbuff* rpcbuff, uint64_t* index, size_t index_len, void* ptype,uint64_t type_len,enum rpctypes type);
 char* rpcbuff_to_arr(struct rpcbuff* rpcbuff,uint64_t* buflen);
 struct rpcbuff* buf_to_rpcbuff(char* buf);
-struct rpcbuff_el* rpcbuff_el_getlast_from(struct  rpcbuff* rpcbuff, uint64_t* index, uint64_t index_len);
+struct rpcbuff_el* rpcbuff_el_getlast_from(struct  rpcbuff* rpcbuff, uint64_t* index, size_t index_len);
 
 
 struct rpcstruct* rpcstruct_create();
